@@ -1,9 +1,8 @@
+import { Logger } from '@map-colonies/js-logger';
 import { RequestHandler } from 'express';
 import httpStatus from 'http-status-codes';
 import { injectable, inject } from 'tsyringe';
 import { Services } from '../../common/constants';
-import { ILogger } from '../../common/interfaces';
-
 import { ChangeManager } from '../models/changeManager';
 import { ChangeWithMetadata, IdMapping } from '../models/types';
 
@@ -22,7 +21,7 @@ export interface MergeChangesRequestBody {
 
 @injectable()
 export class ChangeController {
-  public constructor(@inject(Services.LOGGER) private readonly logger: ILogger, @inject(ChangeManager) private readonly manager: ChangeManager) {}
+  public constructor(@inject(Services.LOGGER) private readonly logger: Logger, @inject(ChangeManager) private readonly manager: ChangeManager) {}
 
   public mergeChanges: MergeChangesHandler = (req, res) => {
     const [change, created, deleted] = this.manager.mergeChanges(req.body.changes, req.body.changesetId);
