@@ -1,16 +1,16 @@
+import { Logger } from '@map-colonies/js-logger';
 import { inject, injectable } from 'tsyringe';
 import { Services } from '../../common/constants';
-import { ILogger } from '../../common/interfaces';
 import { changeToXml } from '../utils/jsonChangeToXml';
 import { mergeChanges } from './merger';
 import { IdMapping, ChangeWithMetadata } from './types';
 
 @injectable()
 export class ChangeManager {
-  public constructor(@inject(Services.LOGGER) private readonly logger: ILogger) {}
+  public constructor(@inject(Services.LOGGER) private readonly logger: Logger) {}
 
   public mergeChanges(changes: ChangeWithMetadata[], changesetId: number): [string, IdMapping[], string[]] {
-    this.logger.log('info', `merging ${changes.length} changes`);
+    this.logger.info(`merging ${changes.length} changes`);
     const [change, idsToCreate, idsToDelete] = mergeChanges(changes, changesetId);
     return [changeToXml({ osmChange: change }), idsToCreate, idsToDelete];
   }
