@@ -1,7 +1,5 @@
-import { ILogMethod } from '@map-colonies/mc-logger';
-
-export interface ILogger {
-  log: ILogMethod;
+export interface IServerConfig {
+  port: string;
 }
 
 export interface IConfig {
