@@ -1,13 +1,20 @@
 import { container } from 'tsyringe';
 import config from 'config';
+import { Tracing, Metrics } from '@map-colonies/telemetry';
+import jsLogger from '@map-colonies/js-logger';
 import { Services } from '../../src/common/constants';
-import { ILogger } from '../../src/common/interfaces';
 
 function registerTestValues(): void {
-  const mockLogger: ILogger = { log: jest.fn() };
-
   container.register(Services.CONFIG, { useValue: config });
-  container.register(Services.LOGGER, { useValue: mockLogger });
+  container.register(Services.LOGGER, { useValue: jsLogger({ enabled: false }) });
+
+  const tracing = new Tracing('app_tracer');
+  const tracer = tracing.start();
+  container.register(Services.TRACER, { useValue: tracer });
+
+  const metrics = new Metrics('app_meter');
+  const meter = metrics.start();
+  container.register(Services.METER, { useValue: meter });
 }
 
 export { registerTestValues };
